backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
feedparser>=6.0.11
pycountry>=23.12.11
sqlalchemy>=2.0.30
apscheduler>=3.10.4
pydantic-settings>=2.3.0
python-dateutil>=2.9.0
